Profile: Remove users name prefixing Meta-analysis
This is in part to fix a formatting issue, if the name of the user is too
long but also follows the same pattern as Github and provides consistency.

Now when viewing a page which is not about you, it will simply omit any
prefix, and when signed in and on a page about you it will prefix with
'Your '.