Papers: prefix doi: and handle in input
There is now a permenant doi: which cannot be removed but does not get
sent to the datastore.

In addition, to better handle copy-pasting into the box the string 'doi:'
is replace with ''. To facilitate this added a new parameter to
addConfirmedUpdater to allow us to reuse the ability to 'trim' a substring.