Papers: Change how adding new comments is handled
This commit changes the behaviour surrounding adding a new comment.

I have used github as inspiration.

Previously onblur of the text field would add it as a comment, and the
'add' button did nothing more but steal focus.

Now I have added a more intuitive behaviour. On focus of the text box
there is a cancel and add button displayed. These will be shown until either
a comment is added, or a user cancels. The add button is disabled if there
is not a valid comment written inside.