Comments: Add edit icon to last comment.
On click focuses the editable text box below it. Ultimately
this just changes the user experience and makes it more obvious
that the comment is editable.

Does not show on non-last comments (non-editable).

Requires material icons script to be included in paper.html to work,
but that is in another branch awaiting merging so omitted here to
save on a merge conflict later.