Metaanalysis: Exclude experiments from aggregate calculations
Checkbox on every experimenttitle, which can be unchecked to add the
experiment into ma.excludedExperiments (in the form <paper.id>,<expIndex>)

Checkbox on every papertitle, which can be unchecked to add all
of that papers experiments into ma.excludedExperiments.

On page load css styling is added to row(s) which was excluded, and
aggregates ignore values if they are from an excluded experiment.