MA/Paper: Display a warning when changing datum values
When the user focuses on a datum which is not empty, change
the class of the changewarning field to re-trigger the animation.

The warning will pulse once from black -> red -> black.

In addition, populate this warning box on page load ONLY if the
datum is not empty, to prevent any wierd popping from focus/unfocus.