Forest plot: Bug fix - Invalid lines problem
When a line was invalid, the totals for the entire graph
did not show. - Change this to use a hasInvalid flag, and
only drop the total for the group(s) containing invalid
total + the final total if any of them are invalid.

Additionally, having any invalid lines caused a display issue,
as we had the +currY line inside the if {show summary} block,
but we need to increment currY regardless.