Graphs: Copy out graph functionality to graphs.js
formulas.js can contain computed cols + aggregates, but graphs
are and should really be something different.

For now, simply copied the functionality out into a new graphs.js
but we should revisit this now that it is factored out, we might not
actually need some of the implementation now we're decoupled from
aggregates.