Graphs: Create basic graph display
Added a graph-table and relevant javascript functions to initially set
and populate it from the MA object.

We do not need a value for graph, since it doesn't calculate anything
for the single value cell, so the value cell has been removed.

The functions still return "see above" but are not called, they will
probably need to be removed if they do not serve a later purpose
of defining the steps to draw the graph.