Metaanalysis: One-click metaanalysis
Introduce the controls and UI for specifying a one-click metaanalysis.

The menus are dynamically drawn in code because HTML templates
wouldn't make anything simpler here.

Three types of option can be drawn:
- Dropdown box: Has an option for each column/aggregate/grouping aggregate
- Manual dropdown box: uses a given list of options
- Checkbox: Can optionally have sub-dropdowns which appear when clicked.

Everything currently is stored in metaanalysis.oneclick **but not saved**.

Additionally, the 'do the magic!' button does not do anything yet.