Metaanalysis: One click add functionality
Add a series of objects to the metaanalysis.

There are sections for compulsory defaults, as well as conditions
for the various options selectable by the user.

Moderator analysis button currently does nothing, awaiting
confirmation on what should be included (although adding grouping
aggregate objects supported).

Ordering is remembered, and a 'thing' wont be added if a similar
object already exists (same formula).