Usernames: version 1
Removes automatic sign up on login, and now redirects to /register.
Adds support for registering a 'username' when a user signs up, and changing it later.
  Checks for username uniqueness and against a list of forbidden words.
URLs/Routes accept either username or email, with a preference to username if it exists.
Username displayed on profile page and button on hover of username links to register page.
Old /api/register renamed to /api/user.
Drop signin.html page entirely.
Redirect lima@local profile to yours if signed in.