Usernames: Messy commit DO NOT PULL
This commit adds to the usernames work, drawing inspiration from
the trello card.

api/storage for the most part is happy to save username, and to work
with a logged in user that they do not know about (some non-breaking errors).

Username displayed and obscure link on hover of username takes to register
page.

Old /api/register renamed to /api/known (although I think we might be able
to remove this entirely).

New /api/register used to save/register a user.