testing: prepare an element for screenshots
Gemini only accepts two options - scroll&stitch everything or nothing, but the latter fails if the element to be screenshot is bigger than the screen.

To work around this, insert an empty element that covers the entire viewport which can be set as a capture element, for use in tests where we only care about the current screen's contents.